Temporarily use simplified Gem installation
This is temporary because it is *nondeterministic*; we're installing multiple Gems in a single command so that RubyGems manages the (varying) compatibility envelope between them. In doing so, we're degrading the future ability (for an unknown-yet-presumably-imminent definition of "future") to replicate the same collection of installed Gems as was done at the time of this commit.

This is because 1.1.1, especially in combination with `hanami-model` 1.1.*0*, tends to produce some dependency incompatibilities when the Gems are installed separately (as our normal procedure would dictate, to install a specific version of each specific Gem).

To the degree that absolute reproducibility is important to you for your use case and shop standards, You Have Been Warned.

To be tagged as 0.12.1.